+++ lib/glimmer/ui/custom_shell.rb
@@ -31,9 +31,9 @@ module Glimmer
       class << self
         def launch(*args, &content)
           auto_exec do
-            @launched_custom_shell = send(keyword, *args, &content)
-            @launched_custom_shell.swt_widget.set_data('launched', true)
-            @launched_custom_shell.open
+            launched_custom_shell = send(keyword, *args, &content)
+            launched_custom_shell.swt_widget.set_data('launched', true)
+            launched_custom_shell.open
           end
         end
       end