+++ samples/elaborate/calculator/model/command/operation/add.rb
@@ -1,0 +1,15 @@
+class Calculator
+  module Model
+    class Command
+      class Operation < Command
+        class Add < Operation
+          keyword '+'
+          
+          def operation_method
+            :+
+          end
+        end
+      end
+    end
+  end
+end